Code restructured into MVC format
